# src/config/swagger.json
{
  "swagger": "2.0",
  "info": {
    "title": "Barefoot Normad API Documentation",
    "version": "1.0.0",
    "description": "Documentation to test the endpoints for the Barefoot Normad API"
  },
  "host": "localhost:4000",
  "basePath": "/api/v1",
...
  "schemes": ["https", "http"],
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "paths": {
...
    "/requests": {
      "post": {
        "tags": ["request"],
        "summary": "Book a travel request",
        "description": "This endpoint books a one-way trip request",
        "operationId": "bookTrip",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "data needed to book a trip",
            "required": true,
            "schema": {
              "$ref": "#/definitions/BookTrip"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "message created successfully",
            "schema": {
              "$ref": "#/definitions/RequestResponse"
            }
          },
          "400": {
            "description": "departureDate is required"
          },
          "409": {
            "description": "You already have a trip slated for this day you may choose to make a multi-city request"
          }
        }
      },
      "get": {
        "tags": ["request"],
        "summary": "a user can view all request he/she made",
        "description": "This end point display all request made by a user",
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "responses": {
          "200": {
            "description": "Trip requests retrieved successfully",
            "schema": {
              "$ref": "#/definitions/RequestResponse"
            }
          },
          "404": {
            "description": "You are yet to book a make a trip request"
          }
        }
      }
    },
...
    "/requests/{requestId}": {
      "delete": {
        "tags": ["Delete request"],
        "summary": "Delete a travel request",
        "description": "This endpoint deletes a single trip request",
        "operationId": "deleteTrip",
        "consumes": ["application/json"],
        "produces": ["application/json"],
        "security": [
          {
            "Bearer": []
          }
        ],
        "parameters": [
          {
            "name": "requestId",
            "in": "path",
            "description": "id of the request to be deleted",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "201": {
            "description": "message deleted successfully",
            "schema": {
              "$ref": "#/definitions/RequestResponse"
            }
          },
          "500": {
            "description": "database error"
          },
          "409": {
            "description": "you already have a trip booked around this period, you may choose to cancel and make a multi-city request"
          }
        }
      }
    },
...
  },
  "definitions": {
...
    "BookTrip": {
      "type": "object",
      "properties": {
        "origin": {
          "type": "string"
        },
        "destination": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "enum": ["one-way", "return"]
        },
        "departureDate": {
          "type": "string",
          "format": "date"
        },
        "returnDate": {
          "type": "string",
          "format": "date"
        },
        "reason": {
          "type": "string"
        },
        "accommodation": {
          "type": "string"
        }
      }
    },
...
    "RequestResponse": {
      "type": "object",
      "properties": {
        "status": {
          "type": "string"
        },
        "data": {
          "type": "object",
          "items": {
            "properties": {
              "origin": {
                "type": "string"
              },
              "destination": {
                "type": "string"
              },
              "type": {
                "type": "string",
                "enum": ["one-way", "return"]
              },
              "departureDate": {
                "type": "string",
                "format": "date"
              },
              "returnDate": {
                "type": "string",
                "format": "date"
              },
              "reason": {
                "type": "string"
              },
              "accommodation": {
                "type": "string"
              }
            }
          }
        }
      }
    }
  }
}